pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.10.0",
    "langchain>=1.2.10",
    "langchain-core>=1.2.16",
    "langgraph>=1.0.10",
    "pillow>=12.1.1",
    "pydot>=4.0.1",
]
